+++ piecrust-uplink/src/lib.rs
@@ -15,8 +15,8 @@ pub use snap::snap;
 
 mod state;
 pub use state::{
-    call, call_raw, caller, height, host_query, limit, meta_data, owner,
-    self_id, spent, ModuleError, State,
+    call, call_raw, caller, emit, height, host_query, limit, meta_data, owner,
+    self_id, spent, ModuleError,
 };
 
 mod helpers;